+++ src/app/modules/step-4-pull-data-page/pull-data-page.module.ts
@@ -15,7 +15,9 @@ import { SearchParameterGroupModule } from '../search-parameter-group/search-par
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { EllipsisTextModule } from '../../shared/ellipsis-text/ellipsis-text.module';
 import { ScrollingModule } from '@angular/cdk/scrolling';
-import { TableVirtualScrollModule } from 'ng-table-virtual-scroll';
+import {
+  TableVirtualScrollModule
+} from '../../shared/table-virtual-scroll/table-virtual-scroll.module';
 
 @NgModule({
   declarations: [PullDataPageComponent],
@@ -36,8 +38,8 @@ import { TableVirtualScrollModule } from 'ng-table-virtual-scroll';
     MatCheckboxModule,
     FormsModule,
     EllipsisTextModule,
-    TableVirtualScrollModule,
-    ScrollingModule
+    ScrollingModule,
+    TableVirtualScrollModule
   ]
 })
 export class PullDataPageModule {}