+++ src/app/shared/table-virtual-scroll/table-virtual-scroll.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {
+  TableVirtualScrollDirective
+} from './table-virtual-scroll.directive';
+
+
+@NgModule({
+  declarations: [TableVirtualScrollDirective],
+  exports: [TableVirtualScrollDirective],
+  imports: [CommonModule]
+})
+export class TableVirtualScrollModule {}
+